parameterize optimizely tests (#94)
* make optimizelytestv2 and optimizelytestv3 as close as possible
* delete OptimizelyTestV2.java
* rename OptimizelyTestV3 to OptimizelyTest
* parameterize Optimizely API Tests based on datafile version
* skip live variable tests for v2 datafile
